services/anomaly-consumer/requirements.txt
cassandra-driver==3.28.0
confluent-kafka==2.3.0